+++ CHANGELOG.md
@@ -36,6 +36,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Changed
 
+- handle jobs in tokio tasks
 - cleaned .env.example and .env.test files
 - bumped snos and downgraded rust to match SNOS rust version
 - Bumped dependencies, and associated api changes done